+++ src/macros.rs
@@ -6,10 +6,10 @@ macro_rules! bounds_check {
             pub (crate) const [<$i:upper _ERROR_NAME>]: &'static str = stringify! { $i };
         }
         paste! {
-            pub (crate) fn [<check_ $i>]($i: u32) -> Result<(), SyroError> {
+            pub (crate) fn [<check_ $i>]($i: u8) -> Result<(), SyroError> {
                 if $i > $hi || $i < $lo {
                     return Err(SyroError::OutOfBounds {
-                        val: $i,
+                        val: $i as u32,
                         name: [<$i:upper _ERROR_NAME>],
                         lo: $lo,
                         hi: $hi,
@@ -27,10 +27,10 @@ macro_rules! max_check {
             pub (crate) const [<$i:upper _ERROR_NAME>]: &'static str = stringify! { $i };
         }
         paste! {
-            pub (crate) fn [<check_ $i>]($i: u32) -> Result<(), SyroError> {
+            pub (crate) fn [<check_ $i>]($i: u8) -> Result<(), SyroError> {
                 if $i > $hi {
                     return Err(SyroError::OutOfBounds {
-                        val: $i,
+                        val: $i as u32,
                         name: [<$i:upper _ERROR_NAME>],
                         lo: 0,
                         hi: $hi,
